+++ game-farfarwest/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
+## [0.3.1] - 2026-06-22
+
+- Fixed: Load Order Context menus no longer cut off when item is at bottom of the application window.
+- Fixed: Scrollbar displayed properly on UE4SS and LogicMods Load Order pages.
+
 ## [0.3.0] - 2026-06-12
 
 - Added: Collections support for UE4SS and LogicMods load orders; exports load orders to the collection manifest, restores them on collection install, and adds a read-only "UE4SS Load Orders" tab to the collection workshop